fix timing advance HW imperfections and ta algorithm in ue

+++ openair2/LAYER2/NR_MAC_COMMON/nr_mac.h
@@ -112,7 +112,7 @@ typedef struct {
   uint8_t R: 2;       // octet 1 [7:6]
 } __attribute__ ((__packed__)) NR_MAC_SUBHEADER_FIXED;
 
-static inline int get_mac_len(uint8_t* pdu, int pdu_len, uint16_t *mac_ce_len, uint16_t *mac_subheader_len) {
+static inline int get_mac_len(uint8_t* pdu, int pdu_len, uint *mac_ce_len, uint *mac_subheader_len) {
   if ( pdu_len < (int)sizeof(NR_MAC_SUBHEADER_SHORT))
     return false;
   NR_MAC_SUBHEADER_SHORT *s = (NR_MAC_SUBHEADER_SHORT*) pdu;
